更新lab3，通过55个点

+++ lab3/src/front/semantic.cpp
@@ -174,7 +174,7 @@ frontend::STE frontend::SymbolTable::get_ste(string id) const {
         if(temp.find(id) != temp.end()){     // 找到了
             return temp[id];
         }
-    }
+    }    
 }
 
 
@@ -899,7 +899,7 @@ void frontend::Analyzer::analysisLVal(LVal* root, vector<ir::Instruction*>& buff
 
         STE arr = symbol_table.get_ste(tk.value);
         vector<int> dimension = arr.dimension;  // 维度
-        int size = dimension.size();    // 数组长度
+        int size = (int)dimension.size();    // 数组长度
 
         // Ident '[' Exp ']'
         if ((int)root->children.size() == 4){     // 一维数组
